Revisions to P9

+++ src/components/Card.js
@@ -18,41 +18,19 @@ export default class Card {
 
   // Event Listeners
   _setEventListeners() {
-    // this._cardElement.addEventListener("click", () => {
-    //   this._cardElement
-    //     .querySelector(".card__like-button")
-    //     .addEventListener("click", () => {
-    //       this.handleLikeClick(this);
-    //     });
-    // });
-
-    // this._cardElement
-    //   .querySelector(".card__delete-button")
-    //   .addEventListener("click", () => {
-    //     this.handleDeleteCard(this._id, this._cardElement);
-    //   });
-
-    // this._cardElement
-    //   .querySelector(".card__image")
-    //   .addEventListener("click", () => {
-    //     this._handleImageClick(this._link, this._name);
-    //   });
     this._cardElement.addEventListener("click", (event) => {
       const target = event.target;
-      if
-      (target.classList.contains("card__like-button")) {
+      if (target.classList.contains("card__like-button")) {
         this.handleLikeClick(this);
       }
-      if
-      (target.classList.contains("card__delete-button")) {
+      if (target.classList.contains("card__delete-button")) {
         this.handleDeleteCard(this._id, this._cardElement);
       }
-      if
-      (target.classList.contains("card__image")) {
+      if (target.classList.contains("card__image")) {
         this.handleImageClick(this._link, this._name);
-    }
+      }
     });
-    }
+  }
 
   toggleLike() {
     this._cardElement
